+++ src/commands/save.rs
@@ -6,16 +6,21 @@ use reqwest::{
     header::{HeaderMap, HeaderName, HeaderValue},
     Proxy, Url,
 };
-use std::{path::PathBuf, sync::Arc};
+use std::{
+    path::{Path, PathBuf},
+    sync::Arc,
+};
 
-/// Download and save HLS and DASH playlists to disk.
+/// Download DASH and HLS playlists.
 #[derive(Debug, Clone, Args)]
 pub struct Save {
-    /// http(s):// | .m3u8 | .m3u | .mpd | .xml
+    /// http(s):// | .mpd | .xml | .m3u8
     #[arg(required = true)]
     pub input: String,
 
-    /// Base url for building segment url. Usually needed for local file.
+    /// Base url to be used for building absolute url to segment.
+    /// This flag is usually needed for local input files.
+    /// By default redirected playlist url is used.
     #[arg(long)]
     pub base_url: Option<Url>,
 
@@ -24,12 +29,9 @@ pub struct Save {
     #[arg(short, long)]
     pub directory: Option<PathBuf>,
 
-    // /// Raw style input prompts for old and unsupported terminals.
-    // #[arg(long)]
-    // pub multiple_progress_bar: bool,
     /// Mux all downloaded streams to a video container (.mp4, .mkv, etc.) using ffmpeg.
     /// Note that existing files will be overwritten and downloaded streams will be deleted.
-    #[arg(short, long, value_parser = output_parser)]
+    #[arg(short, long)]
     pub output: Option<String>,
 
     /// Raw style input prompts for old and unsupported terminals.
@@ -51,6 +53,7 @@ pub struct Save {
     pub prefer_subs_lang: Option<String>,
 
     /// Automatic selection of some standard resolution streams with highest bandwidth stream variant from playlist.
+    /// If matching resolution of WIDTHxHEIGHT is not found then only resolution HEIGHT would be considered for selection.
     /// possible values: [lowest, min, 144p, 240p, 360p, 480p, 720p, hd, 1080p, fhd, 2k, 1440p, qhd, 4k, 8k, highest, max]
     #[arg(short, long, help_heading = "Automation Options", default_value = "highest", value_name = "WIDTHxHEIGHT", value_parser = quality_parser)]
     pub quality: Quality,
@@ -65,42 +68,42 @@ pub struct Save {
 
     /// Custom headers for requests.
     /// This option can be used multiple times.
-    #[arg(long, help_heading = "Client Options", number_of_values = 2, value_names = &["KEY", "VALUE"])]
-    pub header: Vec<String>, // Vec<Vec<String>> not supported
+    #[arg(long, help_heading = "Client Options", num_args = 2, value_names = &["KEY", "VALUE"])]
+    pub header: Vec<String>, // Vec<(String, String)> not supported
 
-    /// Set HTTP(s) / socks proxy for requests.
+    /// Set http(s) / socks proxy address for requests.
     #[arg(long, help_heading = "Client Options", value_parser = proxy_address_parser)]
-    pub proxy_address: Option<Proxy>,
+    pub proxy: Option<Proxy>,
 
     /// Fill request client with some existing cookies per domain.
     /// First value for this option is set-cookie header and second value is url which was requested to send this set-cookie header.
     /// Example `--set-cookie "foo=bar; Domain=yolo.local" https://yolo.local`.
     /// This option can be used multiple times.
-    #[arg(long, help_heading = "Client Options", number_of_values = 2, value_names = &["SET_COOKIE", "URL"])]
-    pub set_cookie: Vec<String>, // Vec<Vec<String>> not supported
+    #[arg(long, help_heading = "Client Options", num_args = 2, value_names = &["SET_COOKIE", "URL"])]
+    pub set_cookie: Vec<String>, // Vec<(String, String)> not supported
 
     /// Update and set user agent header for requests.
     #[arg(
         long,
         help_heading = "Client Options",
-        default_value = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/101.0.4951.64 Safari/537.36"
+        default_value = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36"
     )]
     pub user_agent: String,
 
-    /// Use all supplied keys for decrypting instead of using default kid only.
+    /// Use all supplied keys for decryption instead of using keys which matches with default kid only.
     #[arg(long, help_heading = "Decrypt Options")]
     pub all_keys: bool,
 
-    /// Decryption keys for decrypting CENC encrypted streams.
-    /// Key value should be specified in hex.
-    /// Use `base64:` prefix if key is in base64 format.
-    /// Streams encrypted with a single key can use `--key base64:MhbcGzyxPfkOsp3FS8qPyA==` like key format.
-    /// Streams encrypted with multiple keys can use `--key eb676abbcb345e96bbcf616630f1a3da:100b6c20940f779a4589152b57d2dacb like key format.
+    /// Keys for decrypting encrypted streams.
+    /// If streams are encrypted with a single key then there is no need to specify key id
+    /// else specify decryption key in format KID:KEY.
+    /// KEY value can be specified in hex, base64 or file format.
     /// This option can be used multiple times.
-    #[arg(short, long, help_heading = "Decrypt Options", value_name = "<KID:(base64:)KEY>|(base64:)KEY", value_parser = key_parser)]
+    #[arg(short, long, help_heading = "Decrypt Options", value_name = "KEY|KID:KEY", value_parser = key_parser)]
     pub key: Vec<(Option<String>, String)>,
 
     /// Download encrypted streams without decrypting them.
+    /// Note that --output flag is ignored if this flag is used.
     #[arg(long, help_heading = "Decrypt Options")]
     pub no_decrypt: bool,
 
@@ -171,67 +174,30 @@ fn quality_parser(s: &str) -> Result<Quality, String> {
 }
 
 fn key_parser(s: &str) -> Result<(Option<String>, String), String> {
-    let key = if s.contains(':') && !s.starts_with("base64") {
-        let kid = s.split(':').next().unwrap();
-
-        (
-            Some(kid.to_lowercase().replace('-', "")),
-            s.trim_start_matches(kid)
-                .trim_start_matches(':')
-                .to_string(),
-        )
+    let (key_id, mut key) = if let Some((key_id, key)) = s.split_once(':') {
+        (Some(key_id.to_lowercase().replace('-', "")), key.to_owned())
     } else {
         (None, s.to_owned())
     };
 
-    if key.1.starts_with("base64:") {
-        Ok((
-            key.0,
-            openssl::bn::BigNum::from_slice(
-                &openssl::base64::decode_block(key.1.trim_start_matches("base64:"))
-                    .map_err(|e| e.to_string())?,
-            )
-            .map_err(|e| e.to_string())?
-            .to_hex_str()
-            .map_err(|e| e.to_string())?
-            .to_ascii_lowercase(),
-        ))
+    if let Ok(decoded_key) = openssl::base64::decode_block(&key) {
+        key = hex::encode(decoded_key);
     } else {
-        Ok(key)
-    }
-}
+        let key_file = Path::new(&key);
 
-fn find_ffmpeg() -> Option<String> {
-    Some(
-        std::env::var("PATH")
-            .ok()?
-            .split(if cfg!(target_os = "windows") {
-                ';'
+        if key_file.exists() {
+            if key_file.is_file() {
+                key = hex::encode(
+                    std::fs::read(&key_file)
+                        .map_err(|_| format!("cannot read key from {}", key))?,
+                )
             } else {
-                ':'
-            })
-            .find(|s| {
-                std::path::Path::new(s)
-                    .join(if cfg!(target_os = "windows") {
-                        "ffmpeg.exe"
-                    } else {
-                        "ffmpeg"
-                    })
-                    .exists()
-            })?
-            .to_owned(),
-    )
-}
-
-fn output_parser(s: &str) -> Result<String, String> {
-    if find_ffmpeg().is_some() {
-        Ok(s.to_owned())
-    } else {
-        Err(
-            "could'nt locate ffmpeg binary in PATH (https://www.ffmpeg.org/download.html)"
-                .to_owned(),
-        )
+                return Err("cannot read key from a non file path".to_owned());
+            }
+        }
     }
+
+    Ok((key_id, key.to_lowercase()))
 }
 
 fn proxy_address_parser(s: &str) -> Result<Proxy, String> {
@@ -247,7 +213,7 @@ impl Save {
         if !self.header.is_empty() {
             let mut headers = HeaderMap::new();
 
-            for i in (0..headers.len()).step_by(2) {
+            for i in (0..self.header.len()).step_by(2) {
                 headers.insert(
                     self.header[i].parse::<HeaderName>()?,
                     self.header[i + 1].parse::<HeaderValue>()?,
@@ -257,7 +223,7 @@ impl Save {
             client_builder = client_builder.default_headers(headers);
         }
 
-        if let Some(proxy) = &self.proxy_address {
+        if let Some(proxy) = &self.proxy {
             client_builder = client_builder.proxy(proxy.to_owned());
         }
 
@@ -265,10 +231,8 @@ impl Save {
 
         if !self.set_cookie.is_empty() {
             for i in (0..self.set_cookie.len()).step_by(2) {
-                cookie_jar.add_cookie_str(
-                    &self.set_cookie[i],
-                    &self.set_cookie[i + 1].parse::<reqwest::Url>()?,
-                );
+                cookie_jar
+                    .add_cookie_str(&self.set_cookie[i], &self.set_cookie[i + 1].parse::<Url>()?);
             }
         }