+++ CHANGELOG.md
@@ -28,12 +28,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Also, spinner is removed from progress bar.
 - `--cookies` flag is renamed as `--set-cookie`.
 - `--baseurl` flag is renamed as `--base-url`.
+- `--proxy-address` flag is renamed as `--proxy`.
 - `--quality` flag now also matches height if that specific resolution is not found.
 
 ### Fixed
 
 - Some program panics when auto selecting streams using `--quality` flag.
 - `--directory` flag implementation.
+- `--header` flag implementation.
 - Unknown errors while extracting `stpp` and `application/ttml+xml` streams.
 - Use HLS `#EXT-X-KEY` tag more correctly.
 - DASH stream parsing logic.